requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.2
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=9.5.0
pyyaml>=6.0

# Web and API
flask>=2.3.3
flask-socketio>=5.3.4
flask-cors>=4.0.0
werkzeug>=2.3.7
aiohttp>=3.8.5

# Storage and Caching
redis>=4.6.0
motor>=3.1.1
aioredis>=2.0.1
elasticsearch>=8.8.0
minio>=7.1.15
boto3>=1.28.36
aiofiles>=23.2.1
aio-pika>=9.2.0

# ML and Computer Vision
albumentations>=1.3.1
cloudinary>=1.33.0

# Monitoring and Tracing
prometheus_client>=0.17.1
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-instrumentation-aio-pika>=0.40b0
opentelemetry-exporter-jaeger>=1.19.0
tritonclient>=2.34.0
pynvml>=11.5.0

# Cloud and Deployment
kubernetes>=27.2.0
google-cloud-storage>=2.10.0